more test fixes
