--- temporal-sdk/src/test/java/io/temporal/testUtils/TimingSensitiveTests.java
@@ -1,3 +1,0 @@
-package io.temporal.testUtils;
-
-public interface TimingSensitiveTests {}
